+++ styles/globals.css
@@ -5,7 +5,7 @@
 @tailwind utilities;
 
 *{
-  font-family: 'Roboto', sans-serif;
+  font-family: 'Inter', sans-serif;
   scroll-behavior: smooth;
 }
